% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{add_arc.bnet}
\alias{add_arc.bnet}
\title{add arc to a bnet object}
\usage{
add_arc.bnet(object, from, to, data = NULL, check = T, ...)
}
\arguments{
\item{object}{a bnet obejct}

\item{from}{node to start the arc from}

\item{to}{node to end the arc}

\item{data}{optionally a data.frame for fitting the \code{to} node that
now has an additional parent}

\item{check}{logical, to check or not for acyclic property}

\item{...}{additional parameters to be passed to bmop fitting functions}
}
\value{
a bnet object
}
\description{
add arc to a bnet object
}
\examples{
mat <- rbind(c(0, 0, 1, 1),
c(0, 0, 1, 1),
c(0, 0, 0, 1),
c(0, 0, 0, 0))
bnet<-as.bnet(mat)
plot(bnet)
bnet<-add_arc.bnet(bnet,"1","2")
plot(bnet)
}

