% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{add_nodes.bnet}
\alias{add_nodes.bnet}
\title{add nodes to a bnet object}
\usage{
add_nodes.bnet(object, nodes, data = NULL, ...)
}
\arguments{
\item{object}{a bnet obejct}

\item{nodes}{a vector of names for the nodes to be added}

\item{data}{optionally a data.frame for fitting the new nodes}

\item{...}{additional parameters to be passed to bmop fitting functions}
}
\value{
a bnet object
}
\description{
add nodes to a bnet object
}
\examples{
mat <- rbind(c(0, 0, 1, 1),
c(0, 0, 1, 1),
c(0, 0, 0, 1),
c(0, 0, 0, 0))
bnet<-as.bnet(mat)
plot(bnet)
bnet<-add_nodes.bnet(bnet,"5")
plot(bnet)
}

