% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{adj_matrix.bnet}
\alias{adj_matrix.bnet}
\title{adjacency matrix for a bnet object}
\usage{
adj_matrix.bnet(object)
}
\arguments{
\item{object}{a bnet object}
}
\value{
a matrix, the adjacency matrix relative to \code{object}
}
\description{
adjacency matrix for a bnet object
}

