% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{AIC.bnet}
\alias{AIC.bnet}
\title{AIC of bnet}
\usage{
\method{AIC}{bnet}(object, data, ..., k = 2)
}
\arguments{
\item{object}{a bnet object}

\item{data}{the data to compute the log-likelihood}

\item{...}{compatibility arguments}

\item{k}{penalization coefficient (k=2 standard AIC)}
}
\value{
Numerical, the Akaike Information Criteria computed.
}
\description{
Computes de Akaike Information Criteria for a bnet object over
a dataset of observations
}
\seealso{
\code{\link{BIC.bnet}}
}

