% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{arcs.bnet}
\alias{arcs.bnet}
\title{Arcs of an object}
\usage{
\method{arcs}{bnet}(object)
}
\arguments{
\item{object}{}
}
\value{
a matrix with two columns, each row represent an arc from
the first element to the second.
}
\description{
Arcs of an object
}
\examples{
bnet<-new_bnet(c("A","B"),matrix(nrow=2,c(0,1,0,0)))
plot(bnet)
arcs(bnet)
}

