% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{arcs}
\alias{arcs}
\title{Arcs of an object}
\usage{
arcs(object)
}
\arguments{
\item{object}{}
}
\value{
a matrix with two columns, each row represent an arc from
the first element to the second.
}
\description{
Arcs of an object
}

