% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{as.bnet}
\alias{as.bnet}
\title{Convert object to bnet}
\usage{
as.bnet(x, ...)
}
\arguments{
\item{x}{a R object}

\item{...}{parameter for compatibility reason}
}
\value{
a bnet object
}
\description{
..when possible (matrix, bnlearn, graph)
}
\examples{
mat <- rbind(c(0, 0, 1, 1),
c(0, 0, 1, 1),
c(0, 0, 0, 1),
c(0, 0, 0, 0))
g1 <- graph::graphAM(adjMat=mat,edgemode = "directed")
formula <- Y ~ X1 + X2 + X3 + X4
bnet1 <- as.bnet(g1)
bnet2 <- as.bnet(mat)
bnet3 <- as.bnet(3*mat)
bnet4 <- as.bnet(formula)
}

