% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{as.COSA.bnet}
\alias{as.COSA.bnet}
\title{Convert bnet into list of nodes and arcs}
\usage{
as.COSA.bnet(x)
}
\arguments{
\item{x}{a bnet object}
}
\value{
a list with at least component \code{nodes} and \code{arcs}
}
\description{
Convert bnet into list of nodes and arcs
}

