% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{as.list.bnet}
\alias{as.COSA.bnet}
\alias{as.list.bnet}
\title{Convert bnet into list of nodes and arcs}
\usage{
\method{as.list}{bnet}(x)
}
\arguments{
\item{x}{a bnet object}
}
\value{
a list with at least component \code{nodes} and \code{arcs}
}
\description{
Convert bnet into list of nodes and arcs
}

