% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{BIC.bnet}
\alias{BIC.bnet}
\title{BIC of bnet}
\usage{
\method{BIC}{bnet}(object, data, ...)
}
\arguments{
\item{object}{a bnet object}

\item{data}{the data to compute the log-likelihood}

\item{...}{compatibility arguments}
}
\value{
Numerical, the Bayesian Information Criteria computed.
}
\description{
Computes de Bayesian Information Criteria for a bnet object over
a dataset of observations
}
\seealso{
\code{\link{AIC.bnet}}
}

