% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{bmop_fit.bnet}
\alias{bmop_fit.bnet}
\title{fast fit, internal function}
\usage{
bmop_fit.bnet(object, data, nodes, unif.variables = NULL, Mins = NULL,
  Maxs = NULL, ...)
}
\description{
fast fit, internal function
}

