% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{bmop_mle_fit.bnet}
\alias{bmop_mle_fit.bnet}
\title{very-slow fitting}
\usage{
bmop_mle_fit.bnet(object, data, nodes, unif.variables = NULL, Mins = NULL,
  Maxs = NULL, ...)
}
\description{
very-slow fitting
}

