% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{bmop_search.bnet}
\alias{bmop_search.bnet}
\title{slow-mle fit internal function}
\usage{
bmop_search.bnet(object, data, nodes, unif.variables = NULL, ...)
}
\description{
slow-mle fit internal function
}

