% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{bn_to_bnet}
\alias{bn_to_bnet}
\title{Convert a bnlearn object to bnet object}
\usage{
bn_to_bnet(bn)
}
\arguments{
\item{bn}{a bn object from package bnlearn}
}
\value{
a bnet object
}
\description{
Convert a bnlearn object to bnet object
}
\examples{
\dontrun{data(asia,package= "bnlearn")
bn<-bnlearn::hc(asia)
bnet<-bn_to_bnet(bn)
plot(bnet)
summary(bnet)}
}
\seealso{
bnet_to_bn
}

