% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{bnet_to_bn}
\alias{bnet_to_bn}
\title{Convert a bnet object to bnlearn object}
\usage{
bnet_to_bn(bnet)
}
\arguments{
\item{bnet}{a bnet object}
}
\value{
a bn object from package bnlearn
}
\description{
Convert a bnet object to bnlearn object
}
\seealso{
bn_to_bnet
}

