% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{child.bnet}
\alias{child.bnet}
\title{child of selected nodes in a bnet object}
\usage{
child.bnet(object, nodes)
}
\arguments{
\item{object}{a bnet object}

\item{nodes}{a vector of nodes of bnet}
}
\value{
a vector with the names of the child of \code{nodes}
}
\description{
child of selected nodes in a bnet object
}

