% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{clear_evidence.bnet}
\alias{clear_evidence.bnet}
\title{remove all evidence from a bnet object}
\usage{
clear_evidence.bnet(object)
}
\arguments{
\item{object}{a bnet object}
}
\value{
a bnet object, the same as
\code{object} if old probabilities were not stored when evidence was set,
otherwise the restored bnet object
}
\description{
remove all evidence from a bnet object
}

