% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{collapse.bnet}
\alias{collapse.bnet}
\title{Collapses de bnet removing evidence nodes}
\usage{
collapse.bnet(bn, evidence)
}
\arguments{
\item{bn}{A fitted bnet object}

\item{evidence}{A named list with evidence values}
}
\value{
A bnet
}
\description{
Removes the evidence nodes and update the remaining unsetted nodes
}

