% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{compute_marginal.bnet}
\alias{compute_marginal.bnet}
\title{Marginal Computation}
\usage{
compute_marginal.bnet(object)
}
\arguments{
\item{object}{an bnet object}
}
\value{
an bnet object with marginal computed by integration
}
\description{
Marginal Computation
}

