% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnet-extra-functions.R
\name{create_folds}
\alias{create_folds}
\title{create folds for validation}
\usage{
create_folds(data, k = 10, random = F)
}
\arguments{
\item{data}{a data.frame}

\item{k}{positive integer number of folds}

\item{random,}{logical, random order?}
}
\value{
partitions in folds
}
\description{
create folds for validation
}

