% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{delete_arc.bnet}
\alias{delete_arc.bnet}
\title{delete arc from a bnet object}
\usage{
delete_arc.bnet(object, from, to, data = NULL, ...)
}
\arguments{
\item{object}{a bnet obejct}

\item{from}{node to start the arc from}

\item{to}{node to end the arc}

\item{data}{optionally a data.frame for fitting}

\item{...}{additional parameters to be passed to bmop fitting functions}
}
\value{
a bnet object
}
\description{
delete arc from a bnet object
}
\examples{
mat <- rbind(c(0, 0, 1, 1),
c(0, 0, 1, 1),
c(0, 0, 0, 1),
c(0, 0, 0, 0))
bnet<-as.bnet(mat)
plot(bnet)
bnet<-delete_arc.bnet(bnet,"1","3")
plot(bnet)
}

