% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{delete_node.bnet}
\alias{delete_node.bnet}
\title{delete node (one!) from a bnet object}
\usage{
delete_node.bnet(object, node, data = NULL, mle = F, search = F, ...)
}
\arguments{
\item{object}{a bnet obejct}

\item{node}{node to be deleted}

\item{data}{optionally a data.frame for fitting}

\item{mle}{logical, if use maximum-likelihood estimation in the new nodes}

\item{search}{logical, if use greedy penalized logLik search}

\item{...}{additional parameters to be passed to bmop fitting functions}
}
\value{
a bnet object
}
\description{
delete node (one!) from a bnet object
}
\examples{
mat <- rbind(c(0, 0, 1, 1),
c(0, 0, 1, 1),
c(0, 0, 0, 1),
c(0, 0, 0, 0))
bnet<-as.bnet(mat)
plot(bnet)
bnet<-delete_node.bnet(bnet,"1")
plot(bnet)
}

