% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{densplot.bnetClassifier}
\alias{densplot.bnetClassifier}
\title{Plot various densities for bnet classifier}
\usage{
densplot.bnetClassifier(x, predictor = NULL, ...)
}
\arguments{
\item{x}{a bnetClassifier object}

\item{predictor}{the name of one of the predictor}

\item{...}{other parameters for compatibility}
}
\value{
invisible()
}
\description{
Plot various densities for bnet classifier
}

