% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{dim.bnet}
\alias{dim.bnet}
\title{Dimension of a bnet object}
\usage{
\method{dim}{bnet}(x)
}
\arguments{
\item{x}{a bnet object}
}
\value{
positive number, the dimension of \code{x}
}
\description{
Dimension of a bnet object
}

