% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate an object over points}
\usage{
evaluate(x, object, ...)
}
\arguments{
\item{x}{points over perform evaluations}

\item{object}{R object to be evaluated \code{evaluate.foo} must be available
for object of class \code{foo}}
}
\value{
array of evaluations
}
\description{
This is a generic functions
}

