% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{full_conditioned.bnet}
\alias{full_conditioned.bnet}
\title{Compute univariate conditional distribution}
\usage{
full_conditioned.bnet(bn, node, evidence, normalize = F)
}
\arguments{
\item{bn}{A fitted bnet object}

\item{node}{The name of the node}

\item{evidence}{A named list with values of the MB}

\item{normalize}{Flag. If true the output function is normalized}
}
\value{
List with function, min and max values
}
\description{
Computes the f(x|markovBlanket(x)) function
}

