% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5-bnet-structure-search.R
\name{hillClimb.bnet}
\alias{hillClimb.bnet}
\title{Hill-Climbing structure search}
\usage{
hillClimb.bnet(data, bnet = NULL, whitelist = NULL, blacklist = NULL,
  score = BIC, maxp = dim(data)[2], maxrep = 1000, mle = F,
  search = T, ...)
}
\arguments{
\item{data}{data frame of observation}

\item{bnet}{optional, a bnet object to start the search from}

\item{whitelist}{optional, a matrix or data frame with two columns
       each row indicates an arc}

\item{blacklist}{optional, not implemented}

\item{score}{a score function to be minimized in the search, usually
\code{BIC} or\code{AIC}}

\item{maxp}{positive integer, the maximum number of parents for each node}

\item{maxrep}{positive integer, the maximum number of iterations or jump
       in the search space}

\item{mle}{logical, if use mle estimation of bmop densities}

\item{search}{logical, if use greedy penalized logLik search}

\item{...}{additional parameters to be passed to bmop fitting functions}
}
\value{
A bnet object, the bnet among the ones visited that minimized
        the score.
}
\description{
Function that perform a stocastic hill-climbing search in the space
of possible Bayesian networks, danities and conditional densities are
estimated with a bmop penalized-logLik search, package \code{Rbmop}
}

