% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{is.acyclic}
\alias{is.acyclic}
\title{check if an object fullfil the acyclic property}
\usage{
is.acyclic(object)
}
\arguments{
\item{object}{an R object}
}
\value{
logical
}
\description{
check if an object fullfil the acyclic property
}

