% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{is.bnet}
\alias{is.bnet}
\title{Check if an object is abnet object}
\usage{
is.bnet(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical
}
\description{
Check if an object is abnet object
}

