% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{is.fitted.bnet}
\alias{is.fitted.bnet}
\title{check if a bnet is fitted}
\usage{
is.fitted.bnet(object)
}
\arguments{
\item{object}{a bnet object}
}
\value{
logical
}
\description{
check if a bnet is fitted
}

