% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{marginal.bnet}
\alias{marginal.bnet}
\title{Return marginal x,y to plot}
\usage{
marginal.bnet(node, object, N = 1000, MIN = 0)
}
\arguments{
\item{node}{id of the node to plot the marginal}

\item{object}{bnet object}

\item{N}{number of points}

\item{MIN}{}
}
\description{
Return marginal x,y to plot
}

