% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{markovblanket.bnet}
\alias{markovblanket.bnet}
\title{Markov blanket of a bnet object}
\usage{
markovblanket.bnet(object, nodes)
}
\arguments{
\item{object}{a bnet object}

\item{nodes}{a vector of nodes of \code{object}}
}
\value{
a vector with the markov blanket of the nodes in \code{nodes}
}
\description{
Markov blanket of a bnet object
}
\examples{
\dontrun{data(asia,package="bnlearn")
bn<-bnlearn::hc(asia)
bnet<-bn_to_bnet(bn)
plot(bnet)
summary(bnet)
markovblanket.bnet(bnet,names(asia)[1])}
}

