% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{mcmcSampler.bnet}
\alias{mcmcSampler.bnet}
\title{Generate samples from nodes (with evidence)}
\usage{
mcmcSampler.bnet(bn, evidence, vars = names(bn$variables), N = 1,
  burn.in = 100, thinnin = 10, collapse = F)
}
\arguments{
\item{bn}{A fitted bnet object}

\item{evidence}{A named list with fixed values}

\item{N}{The number of samples to generate}

\item{burn.in}{Number of sampler to ensure chain convergence}

\item{thinnin}{Number of draws to be discarded between samples}
}
\value{
A named list with the complete sample
}
\description{
Generate samples from nodes (with evidence)
}

