% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-bnet-models-functions.R
\name{naive_bayes.bnet}
\alias{naive_bayes.bnet}
\title{Naive Bayes model or BAN model}
\usage{
naive_bayes.bnet(formula = NULL, targets = NULL, predictors = NULL,
  data = NULL, bnet = NULL, ...)
}
\arguments{
\item{formula}{a formula object that specify the model}

\item{targets}{vector of names of the target variables}

\item{predictors}{vector of names of the predictor variables}

\item{data}{optionally a data.frame for bnet fitting}

\item{bnet}{a bnet object, optionally a bnet over the predictor variables}
}
\value{
a Naive Bayes structure, or a BAN model with \code{bnet} as predictor
subgraph
}
\description{
Naive Bayes structure generator for bnet object
}

