% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{new_bnet}
\alias{new_bnet}
\title{New bnet object}
\usage{
new_bnet(variables, adj.mat = NULL)
}
\arguments{
\item{variables}{vector of variables names}

\item{adj.mat}{adjacency matrix, not symmetric, acyclicity will be checked
for every possible arcs.}
}
\value{
a bnet object, that is a list with components $variables,
\code{bnet$variables} is also a list with component named
        as variables, each element represents a node and is composed of
        three elements: \code{name}, \code{parents} and
         \code{prob} (initially set to NULL until fitting of the bnet)
}
\description{
Create a bnet object with prescribed variables and optionally arcs selected
via adjacency matrix.
}
\examples{
bnet1<-new_bnet(c("a","b"))
summary(bnet1)
plot(bnet1)
}

