% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{onevsall_bnet}
\alias{onevsall_bnet}
\title{Classifier One vs All}
\usage{
onevsall_bnet(target, predictors, data = NULL)
}
\arguments{
\item{target}{}

\item{predictors}{}

\item{data}{data set of observations}
}
\value{
a multiclass_bnet object
}
\description{
Classifier One vs All
}

