% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-bnet-manipulate-functions.R
\name{orden.bnet}
\alias{orden.bnet}
\title{Bnet variables ordering}
\usage{
orden.bnet(object)
}
\arguments{
\item{object}{a bnet object}
}
\value{
a bnet object with variables ordered
}
\description{
Bnet variables ordering
}

