% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{plot.bnet}
\alias{plot.bnet}
\title{Plot a bnet object}
\usage{
\method{plot}{bnet}(x, y = NULL, ...)
}
\arguments{
\item{x}{a bnet object}

\item{...}{compatibility with plot.default}
}
\value{
a plot of a bnet structure, with targets (red) and predictors (green)
}
\description{
Plot a bnet object
}
\examples{
bnet<-new_bnet(c("A","B"),matrix(nrow=2,c(0,1,0,0)))
plot(bnet)
}
\seealso{
points.bmop
}

