% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{plot.bnetClassifier}
\alias{plot.bnetClassifier}
\title{Plot a bnet classifier}
\usage{
\method{plot}{bnetClassifier}(x, ...)
}
\arguments{
\item{x}{a bnetClassifier object}

\item{...}{other parameters for compatibility}
}
\value{
invisible()
}
\description{
Plot a bnet classifier
}

