% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{plot_densities.classifier_bnet}
\alias{plot_densities.classifier_bnet}
\title{Plot the densities for every class value}
\usage{
plot_densities.classifier_bnet(object)
}
\arguments{
\item{object}{a classifier_bnet object}
}
\description{
Plot the densities for every class value
}

