% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{predict.bnet}
\alias{predict.bnet}
\title{predict posterior values for a bnet object}
\usage{
\method{predict}{bnet}(object, newdata, targets = NULL, predictors = NULL,
  method = "posteriorMean", optmethod = "Nelder-Mead",
  intMethod = bnetPar()$intMethod, ...)
}
\arguments{
\item{object}{a bnet object}

\item{newdata}{new data, evidence}

\item{targets}{targets variables}

\item{predictors}{predictors variables}

\item{method}{string "posteriorMean" or "posteriorMode"}

\item{optmethod}{\link{optim}}

\item{intmethod}{string "adaptative" or "sampling"}
}
\value{
the values predicted
}
\description{
predict posterior values for a bnet object
}

