% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{predict.bnetClassifier}
\alias{predict.bnetClassifier}
\title{prediction of \code{classifier_bnet} object}
\usage{
\method{predict}{bnetClassifier}(object, newdata, log.odds = F, ...)
}
\description{
prediction of \code{classifier_bnet} object
}

