% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{predict.onevsall_bnet}
\alias{predict.onevsall_bnet}
\title{predict function for onevsall_bnet object}
\usage{
\method{predict}{onevsall_bnet}(object, newdata, ...)
}
\description{
predict function for onevsall_bnet object
}

