% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{print.bnet}
\alias{print.bnet}
\title{Print a bnet object}
\usage{
\method{print}{bnet}(x, ...)
}
\arguments{
\item{x}{a bnet object}

\item{...}{coptaibility arguments}
}
\description{
Print a bnet object
}

