% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{probability.bnet}
\alias{probability.bnet}
\title{compute probability for a bnet}
\usage{
probability.bnet(x, object, log = F)
}
\arguments{
\item{x}{vector of evidence}

\item{object}{}

\item{log}{logical, if \code{TRUE} the logarithm of the probabilities
is returned}
}
\value{
numeric or \code{NULL} if partial evidence is provided
}
\description{
compute probability for a bnet
}

