\name{Rbnet-package}
\alias{Rbnet-package}
\alias{Rbnet}
\docType{package}
\title{
Define and Manage Bayesian Networks
}
\description{
Define and Manage Bayesian Networks, especially with bmop densities estimations
}
\details{
\tabular{ll}{
Package: \tab Rbnet\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-07\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Gherardo Varando <gherardo.varando@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package 
          bayesian network
          rbmop
          b-spline}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
bnet<-new_bnet(c("a","b"))
}
