% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{sample.bnet}
\alias{forwardSampler.bnet}
\alias{sample.bnet}
\title{Generate samples from a bnet}
\usage{
sample.bnet(bn, nodes = variables(bn), evidence = NULL, N = 1,
  burn.in = 100, thinnin = 10, mc.cores = getOption("mc.cores", default =
  4))

forwardSampler.bnet(bn, N = 1, vars = variables(bn))
}
\arguments{
\item{bn}{A fitted bnet object}

\item{evidence}{A named list with fixed values (NULL for no evidence)}

\item{N}{The number of samples to generate}

\item{burn.in}{Number of sampler to ensure chain convergence
(uses \code{mcmc_sampler.bnet})}

\item{thinnin}{Number of draws to be discarded between samples
(uses \code{mcmc_sampler.bnet})}

\item{mc.cores}{Max number of parallel workers}
}
\value{
A data frame with one sample per row (Columns named)
}
\description{
Generate samples from a bnet
}

