% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{summary.bnet}
\alias{summary.bnet}
\title{Summary of a bnet object}
\usage{
\method{summary}{bnet}(object, ...)
}
\arguments{
\item{object}{a bnet object}

\item{...}{compatibility arguments}
}
\description{
Summary of a bnet object
}
\examples{
\dontrun{data(gaussian.test,package="bnlearn")
bn<-bnlearn::hc(gaussian.test)
bnet<-bn_to_bnet(bn)
summary(bnet)}
}

