% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{upper}
\alias{upper}
\title{Upper and Lower values for an object}
\usage{
upper(object)
}
\arguments{
\item{object}{an R object, e.g. a bmop object}
}
\value{
numeric value
}
\description{
This is a generic function, used by package bnet to compute lower and upper
boundaries for the \code{$prob} values of a node.
}

