% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-bnet-evaluation-functions.R
\name{validate.bnet}
\alias{validate.bnet}
\title{validate regression bnet models}
\usage{
validate.bnet(object = NULL, modelfun = NULL, data = NULL,
  targets = NULL, predictors = NULL, folds = NULL, numfolds = 10,
  randomfolds = F, train = NULL, test = NULL, evalfun = MSE, mle = F,
  search = F, mc.cores = 1)
}
\arguments{
\item{object}{optional value, a bnet object}

\item{modelfun}{function to compute a bnet regression model, optional}

\item{data}{data.frame a dataset of observation}

\item{targets}{names of target variables}

\item{predictors}{names of predictor variables}

\item{folds}{folds for validation}

\item{numfolds}{number of folds}

\item{randomfolds}{logical}

\item{train}{indx of train set}

\item{test}{indx of test set}

\item{evalfun}{function, must works as MSE or ARE}

\item{mle}{logica}

\item{mc.cores}{positive integer}
}
\value{
mean and standard deviation of the \code{numfolds} evaluations,
       each one of the evaluations is computed using \code{evalfun}
       (usally \code{MSE}) over the prediction and the true value for the
       given fold (ot train/test set).
}
\description{
validate regression bnet models
}

