% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{variables.bnet}
\alias{variables.bnet}
\title{variables bnet}
\usage{
\method{variables}{bnet}(object)
}
\arguments{
\item{object}{an object \code{bnet}}
}
\value{
vector of string
}
\description{
variables bnet
}

