% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-bnet-classifier.R
\name{variables.bnetClassifier}
\alias{variables.bnetClassifier}
\title{variables for bnetClassifier}
\usage{
\method{variables}{bnetClassifier}(object)
}
\arguments{
\item{object}{a bnetClassiifer object}
}
\value{
vector of variables name
}
\description{
variables for bnetClassifier
}

