% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-bnet-base-functions.R
\name{variables}
\alias{variables}
\title{Variables of an object}
\usage{
variables(object)
}
\arguments{
\item{object}{an R object}
}
\value{
vector of string
}
\description{
This is a generic functions that returns the variables/nodes of an R object.
}
\examples{
bnet1<-new_bnet(c("a","b","c","d"))
variables(bnet1)
}

