% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-bnet-models-functions.R
\name{wrapper}
\alias{wrapper}
\title{wrapper greedy search for structure}
\usage{
wrapper(targets, predictors, data, modelfun = naive_bayes.bnet, bnet = NULL,
  evalfun = "MSE", exit = F)
}
\arguments{
\item{targets}{vector of names for target variables}

\item{predictors}{vector of names for target variables}

\item{data}{data.frame of observations}

\item{modelfun}{function, function that define bnet
must accept data, targets and predictors as input and output a bnet}

\item{bnet}{bnet object, starting strucure or NULL}

\item{evalfun}{string, one of "MSE" or "MAE", or other string
naming a scoring function}

\item{exit}{logica, if \code{TRUE} teh search stops at first not improvement
of the score}
}
\description{
wrapper greedy search based on evaluation of training-error for general funciotn that define a structure
}

