% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-bnet-evaluation-functions.R
\name{zero_one_accuracy}
\alias{zero_one_accuracy}
\title{zero one accuracy}
\usage{
zero_one_accuracy(x, y)
}
\arguments{
\item{x}{a data.frame or a matrix}

\item{y}{a data.frame or a matrix}
}
\value{
positive numbers, the fraction of matching entries between \code{x}
and \code{y}
}
\description{
zero one accuracy
}

