% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{bnetPar}
\alias{bnetPar}
\title{Set fitting parameters}
\usage{
bnetPar(...)
}
\description{
Set appropriate parameters to be used by fitting
functions.
}
\details{
@param ... see details.
 @details This function set parameters used for estimation.
          \code{N}:  The number of knots in every dimensions.
          \code{order}: The order of the B-spline in every dimensions.
          \code{alpha}: The exponent to compute the number of knots.
         \code{knotsMethod}: \code{"uniform"} or \code{"quantiles"} knots.

@export
}

