% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-bnet-models-functions.R
\name{discriminative.bnet}
\alias{discriminative.bnet}
\title{Discriminative model}
\usage{
discriminative.bnet(targets, predictors, data = NULL, ...)
}
\arguments{
\item{targets}{vector of names for targets variables}

\item{predictors}{vector of names for predictor variables}

\item{data}{Training data}
}
\description{
Discriminative model structure generator (inverso of naive bayes)
}

