% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-bnet-structure-search.R
\name{hillClimb.bnet}
\alias{hillClimb.bnet}
\title{Hill-Climbing structure search}
\usage{
hillClimb.bnet(data, bnet = NULL, whitelist = NULL, blacklist = NULL,
  score = BIC, maxp = dim(data)[2], maxrep = 1000, ...)
}
\description{
Function that perform a stocastic hill-climbing search in the space 
of possible Bayesian networks, danities and conditional densities are 
estimated with a bmop penalized-logLik search, package \code{Rbmop}
}
\details{
@param data data frame of observation
 @param bnet optional, a bnet object to start the search from
 @param whitelist optional, a matrix or data frame with two columns 
        each row indicates an arc
 @param blacklist optional, not implemented
 @param score a score function to be minimized in the search, usually 
 \code{BIC} or\code{AIC}
 @param maxp positive integer, the maximum number of parents for each node
 @param maxrep positive integer, the maximum number of iterations or jump
        in the search space
 @param ... additional parameters to be passed to bmop fitting functions
 @return A bnet object, the bnet among the ones visited that minimized 
        the score. 
        @export
}

