% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{lower}
\alias{lower}
\title{This is a generic function, used by package bnet to compute lower and upper
boundaries for the \code{$prob} values of a node.}
\usage{
lower(object)
}
\arguments{
\item{object}{an R object, e.g. a bmop object}
}
\value{
numeric value
}
\description{
This is a generic function, used by package bnet to compute lower and upper
boundaries for the \code{$prob} values of a node.
}

