% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-bnet-fit-inference.R
\name{put_evidence.bnet}
\alias{put_evidence.bnet}
\title{put evidence on a bnet object}
\usage{
put_evidence.bnet(object, evidence, store.old = T, propagate = T)
}
\arguments{
\item{object}{a bnet object}

\item{evidence}{numeric with names or list, the value of evidence}

\item{store.old}{logical}

\item{propagate}{logical, see details}
}
\value{
a bnet object with evidence
}
\description{
put evidence on a bnet object
}
\details{
If  \code{propagate} is set to \code{TRUE}, the evidence will be
naively propagate, that is propagate just to the direct child.
}

